Message the room when combat begins and ends
